refactor(kernel): extract IPC pipe + stream syscalls to kernel/ipc.rs
Phase G.5 of Phase 3 restructure plan.

  - rust/kernel/src/kernel/ipc.rs (new): impl Kernel block over the
    pipe registry (create_pipe / destroy_pipe / close_pipe / has_pipe
    / pipe_write_nowait / pipe_read_nowait / list_pipes /
    pipe_read_blocking / close_all_pipes) and the stream registry
    (create_stream / destroy_stream / close_stream / has_stream /
    stream_write_nowait / stream_read_at / stream_read_batch /
    stream_collect_all / list_streams / stream_read_at_blocking /
    close_all_streams).  ~190 LOC.
  - rust/kernel/src/kernel/mod.rs: drops the corresponding methods
    leaving "moved to" breadcrumbs at both section dividers; declares
    `mod ipc;` alongside `mod dispatch;` `mod locks;`
    `mod observability;`.

Both delegate to PipeManager / StreamManager and persist DT_PIPE /
DT_STREAM inodes via mod.rs's resolve_mount_point + build_metadata
+ commit_metadata helpers — those helpers stay in mod.rs and the
submodule reaches them through `self`.

cargo check -p kernel: clean.
